+++ jest.config.js
@@ -1,4 +1,7 @@
 module.exports = {
   setupFiles: ['<rootDir>/setupTests.js'],
   testPathIgnorePatterns: ['<rootDir>/src/'],
+  transform: {
+    '^.+\\.ts$': 'ts-jest',
+  },
 };